Clang build fix 23

+++ src/WorldStorage/WSSAnvil.cpp
@@ -515,7 +515,7 @@ bool cWSSAnvil::LoadChunkFromNBT(const cChunkCoords & a_Chunk, const cParsedNBT 
 
 			if (block_states_compound > 0)
 			{
-				BlockStateData = GetSectionDataLong(a_NBT, block_states_compound, "data", SectionBlockLongCount);
+				BlockStateData = GetSectionDataLong(a_NBT, block_states_compound, "data", static_cast<size_t>(SectionBlockLongCount));
 				if (BlockStateData != nullptr)
 				{
 					VERIFY(static_cast<size_t>(SectionBlockLongCount) * 8 == a_NBT.GetDataLength(a_NBT.FindChildByName(block_states_compound, "data")));
@@ -523,7 +523,7 @@ bool cWSSAnvil::LoadChunkFromNBT(const cChunkCoords & a_Chunk, const cParsedNBT 
 			}
 			else
 			{
-				BlockStateData = GetSectionDataLong(a_NBT, Child, "BlockStates", SectionBlockLongCount);
+				BlockStateData = GetSectionDataLong(a_NBT, Child, "BlockStates", static_cast<size_t>(SectionBlockLongCount));
 				if (BlockStateData != nullptr)
 				{
 					VERIFY(static_cast<size_t>(SectionBlockLongCount) * 8 == a_NBT.GetDataLength(a_NBT.FindChildByName(Child, "BlockStates")));
@@ -536,7 +536,7 @@ bool cWSSAnvil::LoadChunkFromNBT(const cChunkCoords & a_Chunk, const cParsedNBT 
 			if (BlockStateData != nullptr)
 			{
 				ASSERT(SectionBlockLongCount > 0);
-				UInt64 * LEstates = new UInt64[SectionBlockLongCount];
+				UInt64 * LEstates = new UInt64[static_cast<UInt64>(SectionBlockLongCount)];
 
 				for (size_t i = 0; i < static_cast<size_t>(SectionBlockLongCount); i++)
 				{
@@ -580,14 +580,14 @@ bool cWSSAnvil::LoadChunkFromNBT(const cChunkCoords & a_Chunk, const cParsedNBT 
 					resolveddata[numblockdataindex] = Paletteids[finalv];
 					numblockdataindex++;
 				}
-				Data.BlockData.SetSection(resolveddata, Y);
+				Data.BlockData.SetSection(resolveddata, static_cast<size_t>(Y));
 				delete[] LEstates;
 			}
 			else if (Paletteids.size() == 1)  // if there is only one block in the palette, we can just fill the section with it
 			{
 				std::array<BlockState, 4096> Oneblock;
 				std::fill(Oneblock.begin(), Oneblock.end(), Paletteids[0]);
-				Data.BlockData.SetSection(reinterpret_cast<ChunkBlockData::SectionType &>(Oneblock), Y);
+				Data.BlockData.SetSection(reinterpret_cast<ChunkBlockData::SectionType &>(Oneblock), static_cast<size_t>(Y));
 			}
 			if ((BlockLightData != nullptr) && (SkyLightData != nullptr))
 			{
